Refactor child compiler to integration package
