+++ packages/integration/package.json
@@ -18,6 +18,7 @@
     "@vanilla-extract/css": "^0.4.0",
     "chalk": "^4.1.1",
     "dedent": "^0.7.0",
+    "esbuild": "^0.11.16",
     "eval": "^0.1.6",
     "find-up": "^5.0.0",
     "javascript-stringify": "^2.0.1",
