+++ packages/integration/src/index.ts
@@ -1,4 +1,5 @@
 export { processVanillaFile } from './processVanillaFile';
 export { getSourceFromVirtualCssFile } from './virtualFile';
 export { getPackageInfo } from './packageInfo';
+export { compile, vanillaExtractFilescopePlugin } from './compile';
 export * from './filters';
