+++ packages/vite-plugin/package.json
@@ -19,9 +19,7 @@
   },
   "dependencies": {
     "@vanilla-extract/css": "^0.4.0",
-    "@vanilla-extract/esbuild-plugin": "0.2.0",
-    "@vanilla-extract/integration": "^0.1.0",
-    "esbuild": "^0.11.16"
+    "@vanilla-extract/integration": "^0.1.0"
   },
   "devDependencies": {
     "vite": "^2.2.3"
